+++ include/ag/core/stochastic.hpp
@@ -21,7 +21,7 @@ SampleOut CategoricalSample(const Variable& logits, uint64_t seed=123456789ull);
 
 // Differentiable Gumbel-Softmax (Concrete) relaxation. Returns soft probs of same shape as logits.
 // If 'hard' is true, uses straight-through: hard one-hot in forward, soft in backward.
-Variable GumbelSoftmax(const Variable& logits, double tau=1.0, bool hard=false, uint64_t seed=987654321ull);
+Variable GumbelSoftmax(const Variable& logits, float tau=1.0, bool hard=false, uint64_t seed=987654321ull);
 }
  // namespace ag
