+++ src/ag/core/implicit_qp.cpp
@@ -6,63 +6,63 @@
 namespace ag {
 
 // Tiny dense linear algebra helpers (row-major)
-static void cholesky_spd(std::vector<double>& M, std::size_t n) {
+static void cholesky_spd(std::vector<float>& M, std::size_t n) {
   // In-place Cholesky: M = L*L^T, lower stored in M. Throws if not PD.
   for (std::size_t i=0;i<n;++i) {
     for (std::size_t j=0;j<=i;++j) {
-      double s = M[i*n + j];
+      float s = M[i*n + j];
       for (std::size_t k=0;k<j;++k) s -= M[i*n + k]*M[j*n + k];
       if (i==j) {
-        if (s <= 0.0) throw std::runtime_error("Cholesky failed (matrix not PD)");
+        if (s <= 0.0f) throw std::runtime_error("Cholesky failed (matrix not PD)");
         M[i*n + j] = std::sqrt(s);
       } else {
         M[i*n + j] = s / M[j*n + j];
       }
     }
     // zero upper
-    for (std::size_t j=i+1;j<n;++j) M[i*n + j] = 0.0;
+    for (std::size_t j=i+1;j<n;++j) M[i*n + j] = 0.0f;
   }
 }
-static void chol_solve_lower(const std::vector<double>& L, std::size_t n,
-                             const std::vector<double>& b, std::vector<double>& y) {
-  y.assign(n, 0.0);
+static void chol_solve_lower(const std::vector<float>& L, std::size_t n,
+                             const std::vector<float>& b, std::vector<float>& y) {
+  y.assign(n, 0.0f);
   for (std::size_t i=0;i<n;++i) {
-    double s = b[i];
+    float s = b[i];
     for (std::size_t k=0;k<i;++k) s -= L[i*n + k]*y[k];
     y[i] = s / L[i*n + i];
   }
 }
-static void chol_solve_upper(const std::vector<double>& L, std::size_t n,
-                             const std::vector<double>& y, std::vector<double>& x) {
-  x.assign(n, 0.0);
+static void chol_solve_upper(const std::vector<float>& L, std::size_t n,
+                             const std::vector<float>& y, std::vector<float>& x) {
+  x.assign(n, 0.0f);
   for (int ii=int(n)-1; ii>=0; --ii) {
     std::size_t i = std::size_t(ii);
-    double s = y[i];
+    float s = y[i];
     for (std::size_t k=i+1;k<n;++k) s -= L[k*n + i]*x[k];
     x[i] = s / L[i*n + i];
   }
 }
-static void chol_solve(const std::vector<double>& L, std::size_t n,
-                       const std::vector<double>& b, std::vector<double>& x_out) {
-  std::vector<double> y;
+static void chol_solve(const std::vector<float>& L, std::size_t n,
+                       const std::vector<float>& b, std::vector<float>& x_out) {
+  std::vector<float> y;
   chol_solve_lower(L, n, b, y);
   chol_solve_upper(L, n, y, x_out);
 }
 
 // Compute H^{-1} * v via Cholesky of H (assuming SPD). Reuse factor by passing L factored in H_L if non-empty.
-static void solve_Hinv_times(const std::vector<double>& H_L, std::size_t n,
-                             const std::vector<double>& v, std::vector<double>& x) {
+static void solve_Hinv_times(const std::vector<float>& H_L, std::size_t n,
+                             const std::vector<float>& v, std::vector<float>& x) {
   chol_solve(H_L, n, v, x);
 }
 
 // Build S = A H^{-1} A^T given H factor L (lower)
-static void build_S(const std::vector<double>& H_L, std::size_t n,
-                    const std::vector<double>& A, std::size_t m,
-                    std::vector<double>& S) {
-  S.assign(m*m, 0.0);
+static void build_S(const std::vector<float>& H_L, std::size_t n,
+                    const std::vector<float>& A, std::size_t m,
+                    std::vector<float>& S) {
+  S.assign(m*m, 0.0f);
   // S_ij = e_i^T A H^{-1} A^T e_j = (row i of A) * H^{-1} * (row j of A)^T
   // Compute columns of H^{-1} A^T first
-  std::vector<double> tmp(n), col(n);
+  std::vector<float> tmp(n), col(n);
   for (std::size_t j=0;j<m;++j) {
     // v = A^T e_j  -> take row j of A and treat as v on columns
     for (std::size_t i=0;i<n;++i) tmp[i] = A[j*n + i];
@@ -70,7 +70,7 @@ static void build_S(const std::vector<double>& H_L, std::size_t n,
     solve_Hinv_times(H_L, n, tmp, col);
     // Fill S[:,j] = A * w
     for (std::size_t i=0;i<m;++i) {
-      double s = 0.0;
+      float s = 0.0f;
       for (std::size_t k=0;k<n;++k) s += A[i*n + k] * col[k];
       S[i*m + j] = s;
     }
@@ -78,41 +78,41 @@ static void build_S(const std::vector<double>& H_L, std::size_t n,
 }
 
 // Solve equality-constrained QP via Schur complement
-static void qp_solve_eq_forward(const std::vector<double>& H_in, std::size_t n,
-                                const std::vector<double>& q, 
-                                const std::vector<double>& A, std::size_t m,
-                                const std::vector<double>& b,
-                                std::vector<double>& y_star,
-                                std::vector<double>& lambda_star,
-                                std::vector<double>& H_L_cache) {
+static void qp_solve_eq_forward(const std::vector<float>& H_in, std::size_t n,
+                                const std::vector<float>& q, 
+                                const std::vector<float>& A, std::size_t m,
+                                const std::vector<float>& b,
+                                std::vector<float>& y_star,
+                                std::vector<float>& lambda_star,
+                                std::vector<float>& H_L_cache) {
   // Make H PD (tiny ridge)
-  std::vector<double> H = H_in;
-  for (std::size_t i=0;i<n;++i) H[i*n + i] += 1e-8;
+  std::vector<float> H = H_in;
+  for (std::size_t i=0;i<n;++i) H[i*n + i] += 1e-8f;
   // Cholesky
   cholesky_spd(H, n); // now H is L
   H_L_cache = H;
 
   // Solve for λ: (A H^{-1} A^T) λ = -(b + A H^{-1} q)
-  std::vector<double> S;
+  std::vector<float> S;
   build_S(H_L_cache, n, A, m, S);
 
   // rhs = -(b + A H^{-1} q)
-  std::vector<double> Hinv_q, rhs(m, 0.0);
+  std::vector<float> Hinv_q, rhs(m, 0.0f);
   solve_Hinv_times(H_L_cache, n, q, Hinv_q);
   for (std::size_t i=0;i<m;++i) {
-    double s = -b[i];
+    float s = -b[i];
     for (std::size_t k=0;k<n;++k) s -= A[i*n + k] * Hinv_q[k];
     rhs[i] = s;
   }
   // Solve S λ = rhs (S SPD)
-  std::vector<double> S_L = S;
+  std::vector<float> S_L = S;
   cholesky_spd(S_L, m);
   chol_solve(S_L, m, rhs, lambda_star);
 
   // y = - H^{-1}(q + A^T λ)
-  std::vector<double> tmp(n, 0.0);
+  std::vector<float> tmp(n, 0.0f);
   for (std::size_t i=0;i<n;++i) {
-    double s = q[i];
+    float s = q[i];
     for (std::size_t j=0;j<m;++j) s += A[j*n + i] * lambda_star[j];
     tmp[i] = s;
   }
@@ -122,7 +122,7 @@ static void qp_solve_eq_forward(const std::vector<double>& H_in, std::size_t n,
 
 Variable QPSolveEq(const Variable& H, const Variable& q,
                    const Variable& A, const Variable& b,
-                   double eps_pd) {
+                   float eps_pd) {
   (void)eps_pd; // parameter currently unused; forward adds a tiny ridge internally
 
   // Shapes
@@ -139,13 +139,13 @@ Variable QPSolveEq(const Variable& H, const Variable& q,
 
   auto out = std::make_shared<Node>();
   out->shape = {N};
-  out->value.assign(N, 0.0);
-  out->grad.assign(N, 0.0);
+  out->value.assign(N, 0.0f);
+  out->grad.assign(N, 0.0f);
   out->parents = {H.n, q.n, A.n, b.n};
   out->requires_grad = (H.n->requires_grad || q.n->requires_grad || A.n->requires_grad || b.n->requires_grad);
 
   // forward
-  std::vector<double> y_star, lambda_star, H_L;
+  std::vector<float> y_star, lambda_star, H_L;
   qp_solve_eq_forward(H.n->value, N, q.n->value, A.n->value, M, b.n->value, y_star, lambda_star, H_L);
   out->value = y_star;
 
@@ -162,33 +162,33 @@ Variable QPSolveEq(const Variable& H, const Variable& q,
         (!An || !An->requires_grad) && (!bn || !bn->requires_grad)) return;
 
     // Recompute forward solve (cheap, small dims assumed). Let forward add ridge once.
-    std::vector<double> y_star, lambda_star, H_L;
+    std::vector<float> y_star, lambda_star, H_L;
     qp_solve_eq_forward(Hval, N, qval, Aval, M, bval, y_star, lambda_star, H_L);
 
     // Solve adjoint (KKT)^T * [u_y; u_lambda] = [o->grad; 0]
     // Schur: (A H^{-1} A^T) u_lambda = A H^{-1} bar_y
-    std::vector<double> S;
+    std::vector<float> S;
     build_S(H_L, N, Aval, M, S);
     // rhs_lambda = A H^{-1} bar_y
-    std::vector<double> Hinv_bar_y, rhs_lambda(M, 0.0);
+    std::vector<float> Hinv_bar_y, rhs_lambda(M, 0.0f);
     solve_Hinv_times(H_L, N, o->grad, Hinv_bar_y);
     for (std::size_t i=0;i<M;++i) {
-      double s = 0.0;
+      float s = 0.0f;
       for (std::size_t k=0;k<N;++k) s += Aval[i*N + k] * Hinv_bar_y[k];
       rhs_lambda[i] = s;
     }
     // Solve S u_lambda = rhs
-    std::vector<double> S_L = S, u_lambda;
+    std::vector<float> S_L = S, u_lambda;
     cholesky_spd(S_L, M);
     chol_solve(S_L, M, rhs_lambda, u_lambda);
     // u_y = H^{-1}(bar_y - A^T u_lambda)
-    std::vector<double> tmp(N, 0.0);
+    std::vector<float> tmp(N, 0.0f);
     for (std::size_t i=0;i<N;++i) {
-      double s = o->grad[i];
+      float s = o->grad[i];
       for (std::size_t j=0;j<M;++j) s -= Aval[j*N + i] * u_lambda[j];
       tmp[i] = s;
     }
-    std::vector<double> u_y;
+    std::vector<float> u_y;
     solve_Hinv_times(H_L, N, tmp, u_y);
 
     // Parameter grads (VJP rules)
@@ -200,7 +200,7 @@ Variable QPSolveEq(const Variable& H, const Variable& q,
       // Correct adjoint: dL/dH = -0.5 * (u_y * y_star^T + y_star * u_y^T)
       for (std::size_t i=0;i<N;++i) {
         for (std::size_t j=0;j<N;++j) {
-          double g = -0.5 * (u_y[i]*y_star[j] + y_star[i]*u_y[j]);
+          float g = -0.5f * (u_y[i]*y_star[j] + y_star[i]*u_y[j]);
           Hn->grad[i*N + j] += g;
         }
       }
